refactor(web): extract AI Assistant page logic into modules
Reduce the 1233-line Streamlit page openfatture/web/pages/5_*_AI_Assistant.py to
a thin 78-line wrapper. Extract into openfatture/web/pages/_ai_assistant/:
- helpers.py: handle_chat_error, handle_slash_command (history now an explicit
  param instead of an implicit module global), retry_with_backoff (delegates to
  utils/retry.retry_async with an equivalent RetryConfig, deterministic jitter off)
- tabs.py: render_chat_tab/render_description_tab/render_vat_tab/render_voice_tab
  (pure copy-paste of each tab body)

Also fixes a pre-existing blocker: the page imported get_translator from the
non-existent openfatture.i18n; corrected to openfatture.web.utils.i18n, so the
page actually loads. Adds tests/web/test_ai_assistant_page.py (Streamlit AppTest
smoke + unit tests) — the page had no test before.

Verified: 8/8 tests green, ruff clean, page parses and runs under AppTest.
